backend/requirements.txt
# Provolution Gamification API - Requirements
# Python 3.10+

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
# SQLite is built-in, no extra package needed

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.2

# Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0
email-validator>=2.1.0

# Utilities
python-multipart>=0.0.6  # For form data
python-dotenv>=1.0.0     # Environment variables

# Development
httpx>=0.26.0           # For testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.23.0  # Async test support

# Optional: Production
# gunicorn>=21.2.0       # Production WSGI server
